feat(run.bat): run in Admin mode by default
using a .vbs script (created on the fly), the contents of this bat file
will automatically execute in Admin mode (hopefully everywhere). This
way, chocolatey and ffmpeg can be easily installed.
Change the calling from `call` to `start` as that is required to open a
separate command window after ffmpeg is installed.

Note: Ideally, this whole process should be broken down into two steps.
If the required modules are already installed, just run the script, else
install the requirements first and then run. Currently, requirements are
being installed regardless of whether they are needed or not.